Added gem images to ghosts.
Hopefully also optimized the way structures and gems are handled.
You can now export gems as well. Copying a layout will take a gem as follows: decide the largest color component, then take the displayed gem grade. This means that pure U-upgraded gems are exported as-is, but mixed or weaved gems will be dumbed down. This is so to prevent cheating as I'm not sure how to encode a gem conveniently otherwise.
Copy-pasting within a session still works the same, the gem is duplicated completely.